fix: :bug: ajustes de validações e comentários

+++ database/migrations/2025_05_30_185506_create_produtos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('produtos', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->decimal('preco', 10, 2);
+            $table->integer('preco');
             $table->text('descricao')->nullable();
             $table->string('type');
             $table->timestamps();
